Updates unit tests to account for the refactor
